updates the usage of askID in the tool tips form to be Project ID instead
This replaces the usage of ASK ID in the initial tool tip seed for Project ID. ASK ID is not universally aplicable across organizations, this will allow it to be better deployed to other organizations. This also updates the HTML file to display using projectID instead of askID as a variable name.